#462 - moved filter to app.py
